[Discount svc] Updated implementation for update and delete commands.

+++ src/Services/Discount/Discount.Application/UseCases/Commands/DeleteCoupon/DeleteCouponCommandHandler.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Discount.Application.Exceptions;
 using Discount.Domain.Repositories;
 using MediatR;
 
@@ -10,28 +9,32 @@ namespace Discount.Application.UseCases.Commands.DeleteCoupon
     /// <summary>
     /// The delete coupon command handler.
     /// </summary>
-    public class DeleteCouponCommandHandler : IRequestHandler<DeleteCouponCommand, Unit>
+    public class DeleteCouponCommandHandler : IRequestHandler<DeleteCouponCommand, bool>
     {
         private readonly ICouponRepository _couponRepository;
 
         /// <summary>
         /// Initializes a new instance of a <see cref="DeleteCouponCommandHandler"/>.
         /// </summary>
-        /// <param name="couponRepository"></param>
+        /// <param name="couponRepository">The coupon repository.</param>
         public DeleteCouponCommandHandler(ICouponRepository couponRepository)
         {
             _couponRepository = couponRepository ?? throw new ArgumentNullException(nameof(couponRepository));
         }
 
         /// <inheritdoc />
-        public async Task<Unit> Handle(DeleteCouponCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(DeleteCouponCommand request, CancellationToken cancellationToken)
         {
-            var coupon = await _couponRepository.GetByProductNameAsync(request.ProductName) ??
-                throw new CouponNotFoundException(request.ProductName);
+            var coupon = await _couponRepository.GetByProductNameAsync(request.ProductName);
+
+            if (coupon == null)
+            {
+                return false;
+            }
 
             _couponRepository.Delete(coupon);
 
-            return Unit.Value;
+            return true;
         }
     }
 }
